deinline fmpz_mpoly IO
